+++ library/core/src/main/java/com/google/android/exoplayer2/BaseRenderer.java
@@ -341,6 +341,19 @@ public abstract class BaseRenderer implements Renderer, RendererCapabilities {
    */
   protected final ExoPlaybackException createRendererException(
       Exception cause, @Nullable Format format) {
+    return createRendererException(cause, format, /* isRecoverable= */ false);
+  }
+
+  /**
+   * Creates an {@link ExoPlaybackException} of type {@link ExoPlaybackException#TYPE_RENDERER} for
+   * this renderer.
+   *
+   * @param cause The cause of the exception.
+   * @param format The current format used by the renderer. May be null.
+   * @param isRecoverable If the error is recoverable by disabling and re-enabling the renderer.
+   */
+  protected final ExoPlaybackException createRendererException(
+      Exception cause, @Nullable Format format, boolean isRecoverable) {
     @FormatSupport int formatSupport = RendererCapabilities.FORMAT_HANDLED;
     if (format != null && !throwRendererExceptionIsExecuting) {
       // Prevent recursive re-entry from subclass supportsFormat implementations.
@@ -354,7 +367,7 @@ public abstract class BaseRenderer implements Renderer, RendererCapabilities {
       }
     }
     return ExoPlaybackException.createForRenderer(
-        cause, getName(), getIndex(), format, formatSupport);
+        cause, getName(), getIndex(), format, formatSupport, isRecoverable);
   }
 
   /**