+++ library/core/src/main/java/com/google/android/exoplayer2/ExoPlaybackException.java
@@ -29,9 +29,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.util.concurrent.TimeoutException;
 
-/**
- * Thrown when a non-recoverable playback failure occurs.
- */
+/** Thrown when a non locally recoverable playback failure occurs. */
 public final class ExoPlaybackException extends Exception {
 
   /**
@@ -138,6 +136,17 @@ public final class ExoPlaybackException extends Exception {
    */
   @Nullable public final MediaSource.MediaPeriodId mediaPeriodId;
 
+  /**
+   * Whether the error may be recoverable.
+   *
+   * <p>This is only used internally by ExoPlayer to try to recover from some errors and should not
+   * be used by apps.
+   *
+   * <p>If the {@link #type} is {@link #TYPE_RENDERER}, it may be possible to recover from the error
+   * by disabling and re-enabling the renderers.
+   */
+  /* package */ final boolean isRecoverable;
+
   @Nullable private final Throwable cause;
 
   /**
@@ -167,6 +176,34 @@ public final class ExoPlaybackException extends Exception {
       int rendererIndex,
       @Nullable Format rendererFormat,
       @FormatSupport int rendererFormatSupport) {
+    return createForRenderer(
+        cause,
+        rendererName,
+        rendererIndex,
+        rendererFormat,
+        rendererFormatSupport,
+        /* isRecoverable= */ false);
+  }
+
+  /**
+   * Creates an instance of type {@link #TYPE_RENDERER}.
+   *
+   * @param cause The cause of the failure.
+   * @param rendererIndex The index of the renderer in which the failure occurred.
+   * @param rendererFormat The {@link Format} the renderer was using at the time of the exception,
+   *     or null if the renderer wasn't using a {@link Format}.
+   * @param rendererFormatSupport The {@link FormatSupport} of the renderer for {@code
+   *     rendererFormat}. Ignored if {@code rendererFormat} is null.
+   * @param isRecoverable If the failure can be recovered by disabling and re-enabling the renderer.
+   * @return The created instance.
+   */
+  public static ExoPlaybackException createForRenderer(
+      Exception cause,
+      String rendererName,
+      int rendererIndex,
+      @Nullable Format rendererFormat,
+      @FormatSupport int rendererFormatSupport,
+      boolean isRecoverable) {
     return new ExoPlaybackException(
         TYPE_RENDERER,
         cause,
@@ -175,7 +212,8 @@ public final class ExoPlaybackException extends Exception {
         rendererIndex,
         rendererFormat,
         rendererFormat == null ? RendererCapabilities.FORMAT_HANDLED : rendererFormatSupport,
-        TIMEOUT_OPERATION_UNDEFINED);
+        TIMEOUT_OPERATION_UNDEFINED,
+        isRecoverable);
   }
 
   /**
@@ -225,7 +263,8 @@ public final class ExoPlaybackException extends Exception {
         /* rendererIndex= */ C.INDEX_UNSET,
         /* rendererFormat= */ null,
         /* rendererFormatSupport= */ RendererCapabilities.FORMAT_HANDLED,
-        timeoutOperation);
+        timeoutOperation,
+        /* isRecoverable= */ false);
   }
 
   private ExoPlaybackException(@Type int type, Throwable cause) {
@@ -237,7 +276,8 @@ public final class ExoPlaybackException extends Exception {
         /* rendererIndex= */ C.INDEX_UNSET,
         /* rendererFormat= */ null,
         /* rendererFormatSupport= */ RendererCapabilities.FORMAT_HANDLED,
-        TIMEOUT_OPERATION_UNDEFINED);
+        TIMEOUT_OPERATION_UNDEFINED,
+        /* isRecoverable= */ false);
   }
 
   private ExoPlaybackException(@Type int type, String message) {
@@ -249,7 +289,8 @@ public final class ExoPlaybackException extends Exception {
         /* rendererIndex= */ C.INDEX_UNSET,
         /* rendererFormat= */ null,
         /* rendererFormatSupport= */ RendererCapabilities.FORMAT_HANDLED,
-        /* timeoutOperation= */ TIMEOUT_OPERATION_UNDEFINED);
+        /* timeoutOperation= */ TIMEOUT_OPERATION_UNDEFINED,
+        /* isRecoverable= */ false);
   }
 
   private ExoPlaybackException(
@@ -260,7 +301,8 @@ public final class ExoPlaybackException extends Exception {
       int rendererIndex,
       @Nullable Format rendererFormat,
       @FormatSupport int rendererFormatSupport,
-      @TimeoutOperation int timeoutOperation) {
+      @TimeoutOperation int timeoutOperation,
+      boolean isRecoverable) {
     this(
         deriveMessage(
             type,
@@ -277,7 +319,8 @@ public final class ExoPlaybackException extends Exception {
         rendererFormatSupport,
         /* mediaPeriodId= */ null,
         timeoutOperation,
-        /* timestampMs= */ SystemClock.elapsedRealtime());
+        /* timestampMs= */ SystemClock.elapsedRealtime(),
+        isRecoverable);
   }
 
   private ExoPlaybackException(
@@ -290,7 +333,8 @@ public final class ExoPlaybackException extends Exception {
       @FormatSupport int rendererFormatSupport,
       @Nullable MediaSource.MediaPeriodId mediaPeriodId,
       @TimeoutOperation int timeoutOperation,
-      long timestampMs) {
+      long timestampMs,
+      boolean isRecoverable) {
     super(message, cause);
     this.type = type;
     this.cause = cause;
@@ -301,6 +345,7 @@ public final class ExoPlaybackException extends Exception {
     this.mediaPeriodId = mediaPeriodId;
     this.timeoutOperation = timeoutOperation;
     this.timestampMs = timestampMs;
+    this.isRecoverable = isRecoverable;
   }
 
   /**
@@ -360,7 +405,7 @@ public final class ExoPlaybackException extends Exception {
    * @return The copied exception.
    */
   @CheckResult
-  /* package= */ ExoPlaybackException copyWithMediaPeriodId(
+  /* package */ ExoPlaybackException copyWithMediaPeriodId(
       @Nullable MediaSource.MediaPeriodId mediaPeriodId) {
     return new ExoPlaybackException(
         getMessage(),
@@ -372,7 +417,8 @@ public final class ExoPlaybackException extends Exception {
         rendererFormatSupport,
         mediaPeriodId,
         timeoutOperation,
-        timestampMs);
+        timestampMs,
+        isRecoverable);
   }
 
   @Nullable