+++ library/core/src/main/java/com/google/android/exoplayer2/ExoPlayerImplInternal.java
@@ -142,6 +142,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
   private static final int MSG_PLAYLIST_UPDATE_REQUESTED = 22;
   private static final int MSG_SET_PAUSE_AT_END_OF_WINDOW = 23;
   private static final int MSG_SET_OFFLOAD_SCHEDULING_ENABLED = 24;
+  private static final int MSG_ATTEMPT_ERROR_RECOVERY = 25;
 
   private static final int ACTIVE_INTERVAL_MS = 10;
   private static final int IDLE_INTERVAL_MS = 1000;
@@ -196,6 +197,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
   private long rendererPositionUs;
   private int nextPendingMessageIndexHint;
   private boolean deliverPendingMessageAtStartPositionRequired;
+  @Nullable private ExoPlaybackException pendingRecoverableError;
 
   private long releaseTimeoutMs;
   private boolean throwWhenStuckBuffering;
@@ -525,6 +527,9 @@ import java.util.concurrent.atomic.AtomicBoolean;
         case MSG_SET_OFFLOAD_SCHEDULING_ENABLED:
           setOffloadSchedulingEnabledInternal(msg.arg1 == 1);
           break;
+        case MSG_ATTEMPT_ERROR_RECOVERY:
+          attemptErrorRecovery((ExoPlaybackException) msg.obj);
+          break;
         case MSG_RELEASE:
           releaseInternal();
           // Return immediately to not send playback info updates after release.
@@ -542,9 +547,22 @@ import java.util.concurrent.atomic.AtomicBoolean;
           e = e.copyWithMediaPeriodId(readingPeriod.info.id);
         }
       }
-      Log.e(TAG, "Playback error", e);
-      stopInternal(/* forceResetRenderers= */ true, /* acknowledgeStop= */ false);
-      playbackInfo = playbackInfo.copyWithPlaybackError(e);
+      if (e.isRecoverable && pendingRecoverableError == null) {
+        Log.w(TAG, "Recoverable playback error", e);
+        pendingRecoverableError = e;
+        Message message = handler.obtainMessage(MSG_ATTEMPT_ERROR_RECOVERY, e);
+        // Given that the player is now in an unhandled exception state, the error needs to be
+        // recovered or the player stopped before any other message is handled.
+        message.getTarget().sendMessageAtFrontOfQueue(message);
+      } else {
+        if (pendingRecoverableError != null) {
+          e.addSuppressed(pendingRecoverableError);
+          pendingRecoverableError = null;
+        }
+        Log.e(TAG, "Playback error", e);
+        stopInternal(/* forceResetRenderers= */ true, /* acknowledgeStop= */ false);
+        playbackInfo = playbackInfo.copyWithPlaybackError(e);
+      }
       maybeNotifyPlaybackInfoChanged();
     } catch (IOException e) {
       ExoPlaybackException error = ExoPlaybackException.createForSource(e);
@@ -571,6 +589,19 @@ import java.util.concurrent.atomic.AtomicBoolean;
   }
 
   // Private methods.
+
+  private void attemptErrorRecovery(ExoPlaybackException exceptionToRecoverFrom)
+      throws ExoPlaybackException {
+    Assertions.checkArgument(
+        exceptionToRecoverFrom.isRecoverable
+            && exceptionToRecoverFrom.type == ExoPlaybackException.TYPE_RENDERER);
+    try {
+      seekToCurrentPosition(/* sendDiscontinuity= */ true);
+    } catch (Exception e) {
+      exceptionToRecoverFrom.addSuppressed(e);
+      throw exceptionToRecoverFrom;
+    }
+  }
 
   /**
    * Blocks the current thread until a condition becomes true.
@@ -929,6 +960,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
     } else if (playbackInfo.playbackState == Player.STATE_BUFFERING
         && shouldTransitionToReadyState(renderersAllowPlayback)) {
       setState(Player.STATE_READY);
+      pendingRecoverableError = null; // Any pending error was successfully recovered from.
       if (shouldPlayWhenReady()) {
         startRenderers();
       }
@@ -1318,6 +1350,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
     if (releaseMediaSourceList) {
       mediaSourceList.release();
     }
+    pendingRecoverableError = null;
   }
 
   private Pair<MediaPeriodId, Long> getPlaceholderFirstMediaPeriodPosition(Timeline timeline) {