+++ library/core/src/main/java/com/google/android/exoplayer2/audio/MediaCodecAudioRenderer.java
@@ -37,6 +37,8 @@ import com.google.android.exoplayer2.PlaybackParameters;
 import com.google.android.exoplayer2.PlayerMessage.Target;
 import com.google.android.exoplayer2.RendererCapabilities;
 import com.google.android.exoplayer2.audio.AudioRendererEventListener.EventDispatcher;
+import com.google.android.exoplayer2.audio.AudioSink.InitializationException;
+import com.google.android.exoplayer2.audio.AudioSink.WriteException;
 import com.google.android.exoplayer2.decoder.DecoderInputBuffer;
 import com.google.android.exoplayer2.mediacodec.MediaCodecAdapter;
 import com.google.android.exoplayer2.mediacodec.MediaCodecInfo;
@@ -616,8 +618,10 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
     boolean fullyConsumed;
     try {
       fullyConsumed = audioSink.handleBuffer(buffer, bufferPresentationTimeUs, sampleCount);
-    } catch (AudioSink.InitializationException | AudioSink.WriteException e) {
-      throw createRendererException(e, format);
+    } catch (InitializationException e) {
+      throw createRendererException(e, format, e.isRecoverable);
+    } catch (WriteException e) {
+      throw createRendererException(e, format, e.isRecoverable);
     }
 
     if (fullyConsumed) {
@@ -637,7 +641,8 @@ public class MediaCodecAudioRenderer extends MediaCodecRenderer implements Media
       audioSink.playToEndOfStream();
     } catch (AudioSink.WriteException e) {
       @Nullable Format outputFormat = getOutputFormat();
-      throw createRendererException(e, outputFormat != null ? outputFormat : getInputFormat());
+      throw createRendererException(
+          e, outputFormat != null ? outputFormat : getInputFormat(), e.isRecoverable);
     }
   }